Split store into separate file
